requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.58.2
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
ddgs==9.4.3
distro==1.9.0
duckduckgo_search==8.1.1
ebaysdk==2.2.0
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-anthropic==0.3.17
langchain-community==0.3.27
langchain-core==0.3.69
langchain-ollama==0.3.5
langchain-text-splitters==0.3.8
langgraph==0.6.1
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.1
langgraph-sdk==0.2.0
langsmith==0.4.8
lxml==6.0.0
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.1
ollama==0.5.1
openai==1.97.0
orjson==3.11.0
ormsgpack==1.10.0
outcome==1.3.0.post0
packaging==25.0
primp==0.15.0
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PySocks==1.7.1
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
selenium==4.34.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wikipedia==1.4.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
